load all concept description of one treeItem at once
